# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Goalsy Executive API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface UserProfile {
  userId: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarKey?: string | null;
  onboardingComplete: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateProfileBody {
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarKey?: string | null;
  onboardingComplete?: boolean;
}

export interface FinancialProfile {
  id: string;
  userId: string;
  /** @nullable */
  annualIncome?: number | null;
  /** @nullable */
  monthlyExpenses?: number | null;
  /** @nullable */
  netWorth?: number | null;
  /** @nullable */
  savingsRate?: number | null;
  /** @nullable */
  riskTolerance?: string | null;
  /** @nullable */
  primaryGoalType?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface FinancialProfileResponse {
  profile: FinancialProfile | null;
}

export interface UpdateFinancialProfileBody {
  /** @nullable */
  annualIncome?: number | null;
  /** @nullable */
  monthlyExpenses?: number | null;
  /** @nullable */
  netWorth?: number | null;
  /** @nullable */
  savingsRate?: number | null;
  /** @nullable */
  riskTolerance?: string | null;
  /** @nullable */
  primaryGoalType?: string | null;
}

export interface Goal {
  id: string;
  userId: string;
  name: string;
  type: string;
  targetAmount: number;
  currentAmount: number;
  monthlyContribution: number;
  /** @nullable */
  targetDate?: string | null;
  status: string;
  priority: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateGoalBody {
  name: string;
  type: string;
  targetAmount: number;
  currentAmount?: number;
  monthlyContribution?: number;
  /** @nullable */
  targetDate?: string | null;
  status?: string;
  priority?: number;
}

export interface UpdateGoalBody {
  name?: string;
  type?: string;
  targetAmount?: number;
  currentAmount?: number;
  monthlyContribution?: number;
  /** @nullable */
  targetDate?: string | null;
  status?: string;
  priority?: number;
}

export interface Mission {
  id: string;
  userId: string;
  missionDate: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  status: string;
  /** @nullable */
  skipReason?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface SkipMissionBody {
  reason: string;
}

export interface ScoreDriver {
  label: string;
  value: number;
  maxValue: number;
  trend: string;
}

export interface ScoreResult {
  score: number;
  tier: string;
  drivers: ScoreDriver[];
  computedAt: string;
}

export interface ScoreSnapshot {
  id: string;
  score: number;
  tier: string;
  computedAt: string;
}

export interface NotificationPreference {
  id: string;
  userId: string;
  type: string;
  enabled: boolean;
  updatedAt: string;
}

export interface UpdateNotificationPreferenceBody {
  enabled: boolean;
}

export interface Bill {
  id: string;
  userId: string;
  name: string;
  amount: number;
  dueDate: string;
  /** @nullable */
  account?: string | null;
  /** @nullable */
  category?: string | null;
  isPaid: boolean;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}

export interface Briefing {
  id: string;
  userId: string;
  title: string;
  scheduledDate: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  summary?: string | null;
  createdAt: string;
}

export interface AppNotification {
  id: string;
  userId: string;
  type: string;
  title: string;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  targetScreen?: string | null;
  /** @nullable */
  targetId?: string | null;
  isRead: boolean;
  isDismissed: boolean;
  createdAt: string;
}

/**
 * Not authenticated
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;