# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Goalsy Executive API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /financial-profile:
    get:
      operationId: getFinancialProfile
      tags: [financial-profile]
      summary: Get financial profile (null if not yet set)
      responses:
        "200":
          description: Financial profile or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateFinancialProfile
      tags: [financial-profile]
      summary: Create or update financial profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFinancialProfileBody"
      responses:
        "200":
          description: Saved financial profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List all goals for the current user
      responses:
        "200":
          description: Goals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalBody"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a single goal
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalBody"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /missions/today:
    get:
      operationId: getTodayMission
      tags: [missions]
      summary: Get (or generate) today's mission for the current user
      responses:
        "200":
          description: Today's mission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /missions/{id}/complete:
    post:
      operationId: completeMission
      tags: [missions]
      summary: Mark a mission as completed
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated mission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /missions/{id}/skip:
    post:
      operationId: skipMission
      tags: [missions]
      summary: Skip a mission with a reason
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkipMissionBody"
      responses:
        "200":
          description: Updated mission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /notification-preferences:
    get:
      operationId: listNotificationPreferences
      tags: [notification-preferences]
      summary: List notification preferences
      responses:
        "200":
          description: Preferences list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationPreference"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notification-preferences/{type}:
    put:
      operationId: updateNotificationPreference
      tags: [notification-preferences]
      summary: Enable or disable a notification type
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferenceBody"
      responses:
        "200":
          description: Updated preference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreference"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List bills for the current user
      responses:
        "200":
          description: Bills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /bills/{id}/pay:
    put:
      operationId: payBill
      tags: [bills]
      summary: Mark a bill as paid
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    FinancialProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        annualIncome:
          type: ["integer", "null"]
        monthlyExpenses:
          type: ["integer", "null"]
        netWorth:
          type: ["integer", "null"]
        savingsRate:
          type: ["number", "null"]
        riskTolerance:
          type: ["string", "null"]
        primaryGoalType:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, createdAt, updatedAt]
    FinancialProfileResponse:
      type: object
      properties:
        profile:
          oneOf:
            - $ref: "#/components/schemas/FinancialProfile"
            - type: "null"
      required: [profile]
    UpdateFinancialProfileBody:
      type: object
      properties:
        annualIncome:
          type: ["integer", "null"]
        monthlyExpenses:
          type: ["integer", "null"]
        netWorth:
          type: ["integer", "null"]
        savingsRate:
          type: ["number", "null"]
        riskTolerance:
          type: ["string", "null"]
        primaryGoalType:
          type: ["string", "null"]
    Goal:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        type:
          type: string
        targetAmount:
          type: integer
        currentAmount:
          type: integer
        monthlyContribution:
          type: integer
        targetDate:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, type, targetAmount, currentAmount, monthlyContribution, status, priority, createdAt, updatedAt]
    CreateGoalBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        targetAmount:
          type: integer
        currentAmount:
          type: integer
        monthlyContribution:
          type: integer
        targetDate:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: integer
      required: [name, type, targetAmount]
    UpdateGoalBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        targetAmount:
          type: integer
        currentAmount:
          type: integer
        monthlyContribution:
          type: integer
        targetDate:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: integer
    Mission:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        missionDate:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: string
        skipReason:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, missionDate, title, status, createdAt]
    SkipMissionBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
...
    NotificationPreference:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
        enabled:
          type: boolean
        updatedAt:
          type: string
      required: [id, userId, type, enabled, updatedAt]
    UpdateNotificationPreferenceBody:
      type: object
      properties:
        enabled:
          type: boolean
      required: [enabled]
    Bill:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        amount:
          type: integer
        dueDate:
          type: string
        account:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        isPaid:
          type: boolean
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, name, amount, dueDate, isPaid, createdAt]
...