# lib/api-zod/src/generated/types/bill.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Goalsy Executive API
 * OpenAPI spec version: 0.2.0
 */

export interface Bill {
  id: string;
  userId: string;
  name: string;
  amount: number;
  dueDate: string;
  /** @nullable */
  account?: string | null;
  /** @nullable */
  category?: string | null;
  isPaid: boolean;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}