# lib/api-zod/src/generated/types/createGoalBody.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Goalsy Executive API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateGoalBody {
  name: string;
  type: string;
  targetAmount: number;
  currentAmount?: number;
  monthlyContribution?: number;
  /** @nullable */
  targetDate?: string | null;
  status?: string;
  priority?: number;
}