# lib/api-zod/src/generated/types/financialProfile.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Goalsy Executive API
 * OpenAPI spec version: 0.2.0
 */

export interface FinancialProfile {
  id: string;
  userId: string;
  /** @nullable */
  annualIncome?: number | null;
  /** @nullable */
  monthlyExpenses?: number | null;
  /** @nullable */
  netWorth?: number | null;
  /** @nullable */
  savingsRate?: number | null;
  /** @nullable */
  riskTolerance?: string | null;
  /** @nullable */
  primaryGoalType?: string | null;
  createdAt: string;
  updatedAt: string;
}