# lib/api-zod/src/generated/types/financialProfileResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Goalsy Executive API
 * OpenAPI spec version: 0.2.0
 */
import type { FinancialProfile } from './financialProfile';

export interface FinancialProfileResponse {
  profile: FinancialProfile | null;
}