# lib/api-zod/src/generated/types/notificationPreference.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Goalsy Executive API
 * OpenAPI spec version: 0.2.0
 */

export interface NotificationPreference {
  id: string;
  userId: string;
  type: string;
  enabled: boolean;
  updatedAt: string;
}